feat: migrate validate email route
